Support rank-1 opt for ra-fill!
This can only be done for types #t and 's because bytevector-fill! has
the wrong interface (and even with the right one, it would only work
for 'u8/'s8). Moreover the optimization for #t is hampered by the
implementation of vector-fill! in (srfi :43) which is way slower when
default arguments are provided. We'll have to fix that in Guile.

* bench.scm: Bit larger cases for the benches vs native fill! & copy!.
* mod/newra/map.scm (ra-fill!): Implement the rank-1 opt after the
  pattern in ra-copy!.