Fix bug in rank-1 optimization for ra-copy!
The kernels expected len, but %op-loop-z-1 was passing lenm which is
(- len 1). I was reading the 'end' argument of the -copy! functions
the wrong way and patching it by hand in one too few places, but as it
turns out the 'end' argument was what it should have been.

I spotted the error because array-equal? had a bench time of almost 0.

* test.scm: Add regression.
* bench.scm: Bench ra-equal? with arrays that aren't aren't ra-equal?
* mod/newra/map.scm (%op-loop-z-1): Pass (+ 1 lenm), that is len, to
  kernels. Fix kernels for ra-copy! and ra-fill!.