On array application return element when result has rank 0
This makes array application behave consistently (assuming A is rank 2 as an
example) for (A 0 0) vs (A #%0(0) #%0(0)).

As for ra-from, it can be useful in rank-generic routines that it returns an
array no matter what, so we don't change its behavior, and it's still consistent
in that (ra-from A 0 0) and (ra-from A #%0(0) #%0(0)) both return a rank-0
array.

* mod/newra/base.scm (make-ra*): As stated.
* docs/newra.texi: Examples for array application vs ra-from.
* test/test.scm: Test array application with generalized indices.